@@ -1,105 +1,124 @@
+import Formsy, { addValidationRule, propTypes, withFormsy } from 'formsy-react';
+import PropTypes from 'prop-types';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Formsy from 'formsy-react';
 
 import MyInput from './../components/Input';
 
 const currentYear = new Date().getFullYear();
 
-const validators = {
-  time: {
-    regexp: /^(([0-1]?[0-9])|([2][0-3])):([0-5]?[0-9])(:([0-5]?[0-9]))?$/,
-    message: 'Not valid time'
-  },
-  decimal: {
-    regexp: /(^\d*\.?\d*[0-9]+\d*$)|(^[0-9]+\d*\.\d*$)/,
-    message: 'Please type decimal value'
-  },
-  binary: {
-    regexp: /^([0-1])*$/,
-    message: '10101000'
-  }
-};
-
-Formsy.addValidationRule('isYearOfBirth', (values, value) => {
-  value = parseInt(value);
-  if (typeof value !== 'number') {
+addValidationRule('time', (values, value) => {
+  return /^(([0-1]?[0-9])|([2][0-3])):([0-5]?[0-9])(:([0-5]?[0-9]))?$/.test(value);
+});
+addValidationRule('decimal', (values, value) => {
+  return /(^\d*\.?\d*[0-9]+\d*$)|(^[0-9]+\d*\.\d*$)/.test(value);
+});
+addValidationRule('binary', (values, value) => {
+  return /^([0-1])*$/.test(value);
+});
+addValidationRule('isYearOfBirth', (values, value) => {
+  const parsedValue = parseInt(value, 10);
+  if (typeof parsedValue !== 'number') {
     return false;
   }
-  return value < currentYear && value > currentYear - 130;
+  return parsedValue < currentYear && parsedValue > currentYear - 130;
 });
 
-const App = React.createClass({
+class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.submit = this.submit.bind(this);
+  }
+
   submit(data) {
     alert(JSON.stringify(data, null, 4));
-  },
+  }
+
   render() {
     return (
-      <Formsy.Form onSubmit={this.submit} className="custom-validation">
+      <Formsy onSubmit={this.submit} className="custom-validation">
         <MyInput name="year" title="Year of Birth" type="number" validations="isYearOfBirth" validationError="Please type your year of birth" />
-        <DynamicInput name="dynamic" title="..." />
+        <FormsyDynamicInput name="dynamic" title="..." />
         <button type="submit">Submit</button>
-      </Formsy.Form>
+      </Formsy>
     );
   }
-});
+}
 
-const DynamicInput = React.createClass({
-  mixins: [Formsy.Mixin],
-  getInitialState() {
-    return { validationType: 'time' };
-  },
-  changeValue(event) {
-    this.setValue(event.currentTarget.value);
-  },
+class DynamicInput extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { validationType: 'time' };
+    this.changeValidation = this.changeValidation.bind(this);
+    this.changeValue = this.changeValue.bind(this);
+  }
+
   changeValidation(validationType) {
     this.setState({ validationType: validationType });
-    this.setValue(this.getValue());
-  },
-  validate() {
-    const value = this.getValue();
-    console.log(value, this.state.validationType);
-    return value ? validators[this.state.validationType].regexp.test(value) : true;
-  },
-  getCustomErrorMessage() {
-    return this.showError() ? validators[this.state.validationType].message : '';
-  },
+    this.props.setValue(this.props.getValue());
+  }
+
+  changeValue(event) {
+    this.props.setValue(event.currentTarget.value);
+  }
+
   render() {
-    const className = 'form-group' + (this.props.className || ' ') + (this.showRequired() ? 'required' : this.showError() ? 'error' : null);
-    const errorMessage = this.getCustomErrorMessage();
+    const errorMessage = {
+      time: 'Not a valid time format',
+      decimal: "Not a valid decimal value",
+      binary: "Not a valid binary number"
+    }
 
     return (
-      <div className={className}>
+      <div>
         <label htmlFor={this.props.name}>{this.props.title}</label>
-        <input type='text' name={this.props.name} onChange={this.changeValue} value={this.getValue() || ''}/>
-        <span className='validation-error'>{errorMessage}</span>
-        <Validations validationType={this.state.validationType} changeValidation={this.changeValidation}/>
+        <MyInput validations={this.state.validationType} validationError={errorMessage[this.state.validationType]} type='text' name={this.props.name} onChange={this.changeValue} value={this.props.getValue() || ''} />
+        <Validations validationType={this.state.validationType} changeValidation={this.changeValidation} />
       </div>
     );
   }
-});
+}
 
-const Validations = React.createClass({
+DynamicInput.displayName = "dynamic input";
+
+DynamicInput.propTypes = {
+  ...propTypes,
+};
+
+const FormsyDynamicInput = withFormsy(DynamicInput);
+
+class Validations extends React.Component {
+  constructor(props) {
+    super(props);
+    this.changeValidation = this.changeValidation.bind(this);
+  }
+
   changeValidation(e) {
     this.props.changeValidation(e.target.value);
-  },
+  }
+
   render() {
     const { validationType } = this.props;
     return (
-      <fieldset onChange={this.changeValidation}>
+      <fieldset>
         <legend>Validation Type</legend>
         <div>
-          <input name='validationType' type='radio' value='time' defaultChecked={validationType === 'time'}/>Time
+          <input onChange={this.changeValidation} name='validationType' type='radio' value='time' checked={validationType === 'time'} />Time
         </div>
         <div>
-          <input name='validationType' type='radio' value='decimal' defaultChecked={validationType === 'decimal'}/>Decimal
+          <input onChange={this.changeValidation} name='validationType' type='radio' value='decimal' checked={validationType === 'decimal'} />Decimal
         </div>
         <div>
-          <input name='validationType' type='radio' value='binary' defaultChecked={validationType === 'binary'}/>Binary
+          <input onChange={this.changeValidation} name='validationType' type='radio' value='binary' checked={validationType === 'binary'} />Binary
         </div>
       </fieldset>
     );
   }
-});
+}
 
-ReactDOM.render(<App/>, document.getElementById('example'));
+Validations.propTypes = {
+  changeValidation: PropTypes.func.isRequired,
+  validationType: PropTypes.string.isRequired,
+};
+
+ReactDOM.render(<App />, document.getElementById('example'));
