@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Form } from 'formsy-react';
+import Formsy from 'formsy-react';
 
+import MyCheckbox from './../components/Checkbox';
 import MyInput from './../components/Input';
 import MySelect from './../components/Select';
 
@@ -11,19 +12,45 @@ const user = {
   hair: 'brown'
 };
 
-const App = React.createClass({
+const randomNames = ['Christian', 'Dmitry', 'Aesop'];
+const randomFree = [true, false];
+const randomHair = ['brown', 'black', 'blonde', 'red'];
+
+class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.randomize = this.randomize.bind(this);
+    this.submit = this.submit.bind(this);
+  }
+
+  randomize() {
+    const random = {
+      name: randomNames[Math.floor(Math.random()*randomNames.length)],
+      free: randomFree[Math.floor(Math.random()*randomFree.length)],
+      hair: randomHair[Math.floor(Math.random()*randomHair.length)],
+    };
+
+    this.form.reset(random);
+  }
+
   submit(data) {
     alert(JSON.stringify(data, null, 4));
-  },
-  resetForm() {
-    this.refs.form.reset();
-  },
+  }
+
   render() {
     return (
-      <Formsy.Form ref="form" onSubmit={this.submit} className="form">
+      <Formsy
+        ref={(c) => this.form = c}
+        onSubmit={this.submit}
+        onReset={this.reset}
+        className="form"
+      >
         <MyInput name="name" title="Name" value={user.name} />
-        <MyInput name="free" title="Free to hire" type="checkbox" value={user.free} />
-        <MySelect name="hair" title="Hair" value={user.hair}
+        <MyCheckbox name="free" title="Free to hire" value={user.free} />
+        <MySelect
+          name="hair"
+          title="Hair"
+          value={user.hair}
           options={[
             { value: "black", title: "Black" },
             { value: "brown", title: "Brown" },
@@ -33,12 +60,13 @@ const App = React.createClass({
         />
 
         <div className="buttons">
-          <button type="reset" onClick={this.resetForm}>Reset</button>
+          <button type="reset">Reset</button>
+          <button type="button" onClick={this.randomize}>Randomize</button>
           <button type="submit">Submit</button>
         </div>
-      </Formsy.Form>
+      </Formsy>
     );
   }
-});
+}
 
-ReactDOM.render(<App/>, document.getElementById('example'));
+ReactDOM.render(<App />, document.getElementById('example'));
