@@ -1,10 +1,10 @@
-module.exports = {
-  arraysDiffer: function (a, b) {
-    var isDifferent = false;
+export default {
+  arraysDiffer(a, b) {
+    let isDifferent = false;
     if (a.length !== b.length) {
       isDifferent = true;
     } else {
-      a.forEach(function (item, index) {
+      a.forEach((item, index) => {
         if (!this.isSame(item, b[index])) {
           isDifferent = true;
         }
@@ -13,12 +13,12 @@ module.exports = {
     return isDifferent;
   },
 
-  objectsDiffer: function (a, b) {
-    var isDifferent = false;
+  objectsDiffer(a, b) {
+    let isDifferent = false;
     if (Object.keys(a).length !== Object.keys(b).length) {
       isDifferent = true;
     } else {
-      Object.keys(a).forEach(function (key) {
+      Object.keys(a).forEach((key) => {
         if (!this.isSame(a[key], b[key])) {
           isDifferent = true;
         }
@@ -27,7 +27,7 @@ module.exports = {
     return isDifferent;
   },
 
-  isSame: function (a, b) {
+  isSame(a, b) {
     if (typeof a !== typeof b) {
       return false;
     } else if (Array.isArray(a) && Array.isArray(b)) {
@@ -41,13 +41,61 @@ module.exports = {
     return a === b;
   },
 
-  find: function (collection, fn) {
-    for (var i = 0, l = collection.length; i < l; i++) {
-      var item = collection[i];
+  find(collection, fn) {
+    for (let i = 0, l = collection.length; i < l; i += 1) {
+      const item = collection[i];
       if (fn(item)) {
         return item;
       }
     }
     return null;
-  }
+  },
+
+  runRules(value, currentValues, validations, validationRules) {
+    const results = {
+      errors: [],
+      failed: [],
+      success: [],
+    };
+
+    if (Object.keys(validations).length) {
+      Object.keys(validations).forEach((validationMethod) => {
+        if (validationRules[validationMethod] && typeof validations[validationMethod] === 'function') {
+          throw new Error(`Formsy does not allow you to override default validations: ${validationMethod}`);
+        }
+
+        if (!validationRules[validationMethod] && typeof validations[validationMethod] !== 'function') {
+          throw new Error(`Formsy does not have the validation rule: ${validationMethod}`);
+        }
+
+        if (typeof validations[validationMethod] === 'function') {
+          const validation = validations[validationMethod](currentValues, value);
+          if (typeof validation === 'string') {
+            results.errors.push(validation);
+            results.failed.push(validationMethod);
+          } else if (!validation) {
+            results.failed.push(validationMethod);
+          }
+          return;
+        } else if (typeof validations[validationMethod] !== 'function') {
+          const validation = validationRules[validationMethod](
+            currentValues, value, validations[validationMethod],
+          );
+          if (typeof validation === 'string') {
+            results.errors.push(validation);
+            results.failed.push(validationMethod);
+          } else if (!validation) {
+            results.failed.push(validationMethod);
+          } else {
+            results.success.push(validationMethod);
+          }
+          return;
+        }
+
+        results.success.push(validationMethod);
+      });
+    }
+
+    return results;
+  },
 };
