@@ -1,78 +1,73 @@
-var isExisty = function (value) {
-  return value !== null && value !== undefined;
-};
+const isExisty = value => value !== null && value !== undefined;
+const isEmpty = value => value === '';
 
-var isEmpty = function (value) {
-  return value === '';
-};
-
-var validations = {
-  isDefaultRequiredValue: function (values, value) {
+const validations = {
+  isDefaultRequiredValue(values, value) {
     return value === undefined || value === '';
   },
-  isExisty: function (values, value) {
+  isExisty(values, value) {
     return isExisty(value);
   },
-  matchRegexp: function (values, value, regexp) {
+  matchRegexp(values, value, regexp) {
     return !isExisty(value) || isEmpty(value) || regexp.test(value);
   },
-  isUndefined: function (values, value) {
+  isUndefined(values, value) {
     return value === undefined;
   },
-  isEmptyString: function (values, value) {
+  isEmptyString(values, value) {
     return isEmpty(value);
   },
-  isEmail: function (values, value) {
-    return validations.matchRegexp(values, value, /^((([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+(\.([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+)*)|((\x22)((((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(([\x01-\x08\x0b\x0c\x0e-\x1f\x7f]|\x21|[\x23-\x5b]|[\x5d-\x7e]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(\\([\x01-\x09\x0b\x0c\x0d-\x7f]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]))))*(((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(\x22)))@((([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.)+(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))$/i);
+  isEmail(values, value) {
+    return validations.matchRegexp(values, value, /^([\w-]+(?:\.[\w-]+)*)@((?:[\w-]+\.)*\w[\w-]{0,66})\.([a-z]{2,6}(?:\.[a-z]{2})?)$/i);
   },
-  isUrl: function (values, value) {
-    return validations.matchRegexp(values, value, /^(https?|s?ftp):\/\/(((([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(%[\da-f]{2})|[!\$&'\(\)\*\+,;=]|:)*@)?(((\d|[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.(\d|[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.(\d|[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.(\d|[1-9]\d|1\d\d|2[0-4]\d|25[0-5]))|((([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.)+(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.?)(:\d*)?)(\/((([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(%[\da-f]{2})|[!\$&'\(\)\*\+,;=]|:|@)+(\/(([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(%[\da-f]{2})|[!\$&'\(\)\*\+,;=]|:|@)*)*)?)?(\?((([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(%[\da-f]{2})|[!\$&'\(\)\*\+,;=]|:|@)|[\uE000-\uF8FF]|\/|\?)*)?(#((([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(%[\da-f]{2})|[!\$&'\(\)\*\+,;=]|:|@)|\/|\?)*)?$/i);
+  isUrl(values, value) {
+    return validations.matchRegexp(values, value, /^(?:\w+:)?\/\/([^\s.]+\.\S{2}|localhost[:?\d]*)\S*$/i);
   },
-  isTrue: function (values, value) {
+  isTrue(values, value) {
     return value === true;
   },
-  isFalse: function (values, value) {
+  isFalse(values, value) {
     return value === false;
   },
-  isNumeric: function (values, value) {
+  isNumeric(values, value) {
     if (typeof value === 'number') {
       return true;
     }
     return validations.matchRegexp(values, value, /^[-+]?(?:\d*[.])?\d+$/);
   },
-  isAlpha: function (values, value) {
+  isAlpha(values, value) {
     return validations.matchRegexp(values, value, /^[A-Z]+$/i);
   },
-  isAlphanumeric: function (values, value) {
+  isAlphanumeric(values, value) {
     return validations.matchRegexp(values, value, /^[0-9A-Z]+$/i);
   },
-  isInt: function (values, value) {
+  isInt(values, value) {
     return validations.matchRegexp(values, value, /^(?:[-+]?(?:0|[1-9]\d*))$/);
   },
-  isFloat: function (values, value) {
-    return validations.matchRegexp(values, value, /^(?:[-+]?(?:\d+))?(?:\.\d*)?(?:[eE][\+\-]?(?:\d+))?$/);
+  isFloat(values, value) {
+    return validations.matchRegexp(values, value, /^(?:[-+]?(?:\d+))?(?:\.\d*)?(?:[eE][+-]?(?:\d+))?$/);
   },
-  isWords: function (values, value) {
+  isWords(values, value) {
     return validations.matchRegexp(values, value, /^[A-Z\s]+$/i);
   },
-  isSpecialWords: function (values, value) {
+  isSpecialWords(values, value) {
     return validations.matchRegexp(values, value, /^[A-Z\s\u00C0-\u017F]+$/i);
   },
-  isLength: function (values, value, length) {
+  isLength(values, value, length) {
     return !isExisty(value) || isEmpty(value) || value.length === length;
   },
-  equals: function (values, value, eql) {
-    return !isExisty(value) || isEmpty(value) || value == eql;
+  equals(values, value, eql) {
+    return !isExisty(value) || isEmpty(value) || value === eql;
   },
-  equalsField: function (values, value, field) {
-    return value == values[field];
+  equalsField(values, value, field) {
+    return value === values[field];
   },
-  maxLength: function (values, value, length) {
+  maxLength(values, value, length) {
     return !isExisty(value) || value.length <= length;
   },
-  minLength: function (values, value, length) {
+  minLength(values, value, length) {
     return !isExisty(value) || isEmpty(value) || value.length >= length;
-  }
+  },
 };
 
-module.exports = validations;
+export default validations;
