@@ -1,24 +1,24 @@
 import React from 'react';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 
 import Formsy from './..';
 import { InputFactory } from './utils/TestInput';
 
 const TestInput = InputFactory({
   render() {
-    return <input value={this.getValue()} readOnly/>;
+    return <input value={this.props.getValue()} readOnly/>;
   }
 });
 
-const TestForm = React.createClass({
+class TestForm extends React.Component {
   render() {
     return (
-      <Formsy.Form>
+      <Formsy>
         <TestInput name="foo" validations={this.props.rule} value={this.props.inputValue}/>
-      </Formsy.Form>
+      </Formsy>
     );
   }
-});
+}
 
 export default {
 
