@@ -1,21 +1,25 @@
 import React from 'react';
-import Formsy from './../..';
+import Formsy, { withFormsy } from './../..';
 
-const defaultProps = {
-  mixins: [Formsy.Mixin],
-  getDefaultProps() {
-    return { type: 'text' };
-  },
-  updateValue(event) {
-    this.setValue(event.target[this.props.type === 'checkbox' ? 'checked' : 'value']);
-  },
-  render() {
-    return <input type={this.props.type} value={this.getValue()} onChange={this.updateValue}/>;
-  }
-};
+class TestInput extends React.Component {
+    static defaultProps = { type: 'text' };
 
-export function InputFactory(props) {
-  return React.createClass(Object.assign(defaultProps, props));
+    updateValue = (event) => {
+        this.props.setValue(event.target[this.props.type === 'checkbox' ? 'checked' : 'value']);
+    }
+
+    render() {
+        return <input type={this.props.type} value={this.props.getValue()} onChange={this.updateValue}/>;
+    }
 }
 
-export default React.createClass(defaultProps);
+export function InputFactory(methods) {
+    for (let method in methods) {
+        if (methods.hasOwnProperty(method)) {
+            TestInput.prototype[method] = methods[method];
+        }
+    }
+    return withFormsy(TestInput);
+}
+
+export default withFormsy(TestInput);
